refactor: 파일 도메인에서 팀 도메인으로 이동
