+++ src/main/java/com/ops/ops/modules/team/domain/dao/TeamRepository.java
@@ -3,8 +3,6 @@ package com.ops.ops.modules.team.domain.dao;
 import com.ops.ops.modules.team.domain.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface TeamRepository extends JpaRepository<Team, Long> {
-    Optional<Team> findByTeamIdAndIsDeletedFalse(Long teamId);
+//    Optional<Team> findByTeamIdAndIsDeletedFalse(Long teamId);
 }
